Refactoring utils, adding export for readRemoteReg, writeRemoteReg and readRemoteBlock
