+++ include/ctp7_modules/common/utils.h
@@ -9,9 +9,6 @@
 
 #include "xhal/common/rpc/common.h"
 
-#include <log4cplus/logger.h>
-#include <log4cplus/loggingmacros.h>
-
 #include <ostream>
 #include <iomanip>
 #include <string>
@@ -46,65 +43,6 @@ namespace utils {
     };
 
     /*!
-     *  \brief Object holding counters of errors encountered during VFAT slow-control transactions
-     */
-    struct SlowCtrlErrCntVFAT
-    {
-      uint32_t crc;           ///< GEM_AMC.SLOW_CONTROL.VFAT3.CRC_ERROR_CNT
-      uint32_t packet;        ///< GEM_AMC.SLOW_CONTROL.VFAT3.PACKET_ERROR_CNT
-      uint32_t bitstuffing;   ///< GEM_AMC.SLOW_CONTROL.VFAT3.BITSTUFFING_ERROR_CNT
-      uint32_t timeout;       ///< GEM_AMC.SLOW_CONTROL.VFAT3.TIMEOUT_ERROR_CNT
-      uint32_t axi_strobe;    ///< GEM_AMC.SLOW_CONTROL.VFAT3.AXI_STROBE_ERROR_CNT
-      uint32_t sum;           ///< Sum of above counters
-      uint32_t nTransactions; ///< GEM_AMC.SLOW_CONTROL.VFAT3.TRANSACTION_CNT
-
-      SlowCtrlErrCntVFAT() {
-        crc = packet = bitstuffing = timeout = axi_strobe = sum = nTransactions = 0;
-      }
-
-      SlowCtrlErrCntVFAT(uint32_t crc, uint32_t packet, uint32_t bitstuffing, uint32_t timeout, uint32_t axi_strobe, uint32_t sum, unsigned nTransactions) :
-      crc(crc), packet(packet), bitstuffing(bitstuffing), timeout(timeout), axi_strobe(axi_strobe), sum(sum), nTransactions(nTransactions)
-      {}
-
-      SlowCtrlErrCntVFAT operator+(const SlowCtrlErrCntVFAT &vfatErrs) const {
-        return SlowCtrlErrCntVFAT(crc+vfatErrs.crc,
-                                  packet+vfatErrs.packet,
-                                  bitstuffing+vfatErrs.bitstuffing,
-                                  timeout+vfatErrs.timeout,
-                                  axi_strobe+vfatErrs.axi_strobe,
-                                  sum+vfatErrs.sum,
-                                  nTransactions+vfatErrs.nTransactions);
-      }
-
-      /*!
-       *  \brief Function to detect if an overflow occurs during an addition operation
-       *         (adapted from https://stackoverflow.com/questions/199333/how-do-i-detect-unsigned-integer-multiply-overflow)
-       *
-       *  \param \c a number to be added to \c b
-       *  \param \c b number to add to \c a
-       *
-       *  \returns \c 0xffffffff in the case an overflow is detected, otherwise \c a+b
-       *
-       */
-      inline uint32_t overflowTest(const uint32_t &a, const uint32_t &b) {
-        const uint32_t overflowTest = a+b;
-        if ((overflowTest-b) != a) {
-          return 0xffffffff;
-        } else {
-          return a+b;
-        }
-      }
-
-      void sumErrors() {
-        sum = overflowTest(sum, crc);
-        sum = overflowTest(sum, packet);
-        sum = overflowTest(sum, bitstuffing);
-        sum = overflowTest(sum, timeout);
-        sum = overflowTest(sum, axi_strobe);
-      }
-    };
-
-    /*!
      *  \brief Updates the LMDB object
      *
      *  \param \c at_xml is the name of the XML address table to use to populate the LMDB
@@ -124,6 +62,45 @@ namespace utils {
     struct readRegFromDB : public xhal::common::rpc::Method
     {
         RegInfo operator()(const std::string &regName) const;
+    };
+
+    /*!
+     *  \brief Reads a value from remote register. Exported method. Register mask is applied. An \c std::runtime_error is thrown if the register cannot be read.
+     *
+     *  \param \c regName Register name
+     *
+     *  \returns \c uint32_t register value
+     */
+    struct readRemoteReg : public xhal::common::rpc::Method
+    {
+        uint32_t operator()(const std::string &regName) const;
+    };
+
+    /*!
+     *  \brief Reads a block of values from a contiguous address space.
+     *
+     *  \param \c regName Register name of the block to be read
+     *  \param \c size number of words to read (should this just come from the register properties?
+     *  \param \c offset Start reading from an offset from the base address returned by regName
+     *
+     *  \returns the vector of uint32_t words in the result 
+     */
+
+    struct readRemoteBlock : public xhal::common::rpc::Method
+    {
+        std::vector<uint32_t> operator()(const std::string &regName, const uint32_t size, const uint32_t offset=0) const;
+    };
+
+    /*!
+     *  \brief Writes a value to remote register. Exported method. Register mask is applied. An \c std::runtime_error is thrown if the register cannot be read.
+     *
+     *  \param \c regName Register name
+     *
+     *  \param \c value Value to write
+     */
+    struct writeRemoteReg : public xhal::common::rpc::Method
+    {
+        void operator()(const std::string &regName, const uint32_t value) const;
     };
 }
 #endif
