+++ src/extras.cpp
@@ -5,6 +5,7 @@
 
 #include "ctp7_modules/server/memhub.h"
 #include "ctp7_modules/common/utils.h"
+#include "ctp7_modules/server/utils.h"
 
 #include <libmemsvc.h>
